feat: implement controllers and services for chat, guide, and session management with error handling
