+++ src/services/chatService.ts
@@ -1,0 +1,5 @@
+import { AppDataSource } from "../config/database";
+
+export class ChatService {
+  // TODO: Implement ChatService
+}
